Divide project onto packages
