--- kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/InformationService.java
@@ -1,5 +1,0 @@
-package com.kodilla.good.patterns.challenges;
-
-public interface InformationService {
-    void inform(User user);
-}
