--- kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/Order.java
@@ -1,7 +1,0 @@
-package com.kodilla.good.patterns.challenges;
-
-public interface Order {
-    User getUser();
-    Product getProduct();
-    int getQuantity();
-}
