+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/challenge1/MovieStore.java
@@ -1,4 +1,4 @@
-package com.kodilla.good.patterns.challenges;
+package com.kodilla.good.patterns.challenges.challenge1;
 
 import java.util.ArrayList;
 import java.util.HashMap;
