+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/challenge2/orders/OrderProcessor.java
@@ -1,4 +1,7 @@
-package com.kodilla.good.patterns.challenges;
+package com.kodilla.good.patterns.challenges.challenge2.orders;
+
+import com.kodilla.good.patterns.challenges.challenge2.services.InformationService;
+import com.kodilla.good.patterns.challenges.challenge2.services.ProductOrderService;
 
 import java.time.LocalDate;
