+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/challenge2/products/Product.java
@@ -1,4 +1,4 @@
-package com.kodilla.good.patterns.challenges;
+package com.kodilla.good.patterns.challenges.challenge2.products;
 
 public interface Product {
