+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/challenge2/services/InformationService.java
@@ -1,0 +1,7 @@
+package com.kodilla.good.patterns.challenges.challenge2.services;
+
+import com.kodilla.good.patterns.challenges.challenge2.users.User;
+
+public interface InformationService {
+    void inform(User user);
+}
