+++ kodilla-good-patterns/src/main/java/com/kodilla/good/patterns/challenges/challenge2/users/User.java
@@ -1,4 +1,4 @@
-package com.kodilla.good.patterns.challenges;
+package com.kodilla.good.patterns.challenges.challenge2.users;
 
 public class User {
